requirements.txt
torch
opencv-python
numpy
pathlib
torchvision
Pillow